added rename_fouvars() function including testcase, removed deprecated example script
