+++ tests/test_dfm_tools.py
@@ -89,6 +89,30 @@ def test_rename_waqvars():
     assert 'mesh2d_Chlfa' in uds.data_vars
 
 
+@pytest.mark.requiresdata
+@pytest.mark.unittest
+def test_rename_fouvars_regular():
+    file_nc_fou = r'p:\archivedprojects\11203379-005-mwra-updated-bem\03_model\02_final\A72_ntsu0_kzlb2\DFM_OUTPUT_MB_02_fou\MB_02_0000_fou.nc'
+    uds = dfmt.open_partitioned_dataset(file_nc_fou)
+    uds_renamed = dfmt.rename_fouvars(uds)
+    
+    assert 'mesh2d_tem_mean_20160201000000_20160301000000' in uds_renamed.data_vars
+    assert 'mesh2d_uy_mean_20160101000000_20170101000000' in uds_renamed.data_vars
+    
+    
+@pytest.mark.requiresdata
+@pytest.mark.unittest
+def test_rename_fouvars_tidal():
+    file_nc_fou = r'p:\1230882-emodnet_hrsm\GTSMv3.0EMODnet\EMOD_MichaelTUM_yearcomponents\GTSMv4.1_yeartide_2014_2.20.06\output\gtsm_model_0000_fou.nc'
+    uds = dfmt.open_partitioned_dataset(file_nc_fou)
+    uds_renamed = dfmt.rename_fouvars(uds)
+    
+    assert 'mesh2d_s1_mean_20131231000000_20150101000000' in uds_renamed.data_vars
+    assert 'mesh2d_s1_min_20131231000000_20150101000000' in uds_renamed.data_vars
+    assert 'mesh2d_s1_mindepth_20131231000000_20150101000000' in uds_renamed.data_vars
+    assert 'mesh2d_s1_ampM2_20131231000000_20141227180000' in uds_renamed.data_vars
+
+
 @pytest.mark.systemtest
 def test_cartopy_epsg():
     #this one crashes if the dummy in plot_background() is not created
